Refactor moving logic to connector to fix tests.
